+++ src/test/java/io/github/astrapi69/swing/table/model/dynamic/PermissionsDynamicTableModel.java
@@ -27,7 +27,7 @@ package io.github.astrapi69.swing.table.model.dynamic;
 import java.util.Optional;
 
 import lombok.NonNull;
-import io.github.astrapi69.test.objects.Permission;
+import io.github.astrapi69.test.object.Permission;
 
 /**
  * The class PermissionsDynamicTableModel