+++ src/test/java/io/github/astrapi69/swing/table/panel/DemoTablePanel.java
@@ -33,7 +33,7 @@ import io.github.astrapi69.model.GenericModel;
 import io.github.astrapi69.model.api.IModel;
 import io.github.astrapi69.swing.table.model.DemoPermissionsTableModel;
 import io.github.astrapi69.swing.table.model.GenericTableModel;
-import io.github.astrapi69.test.objects.Permission;
+import io.github.astrapi69.test.object.Permission;
 import io.github.astrapi69.window.adapter.CloseWindow;
 
 public class DemoTablePanel extends TablePanel<Permission>