+++ src/test/java/io/github/astrapi69/swing/table/panel/shuffle/TestAbstractShuffleTablePanel.java
@@ -31,8 +31,8 @@ import javax.swing.JFrame;
 import io.github.astrapi69.model.GenericModel;
 import io.github.astrapi69.swing.table.model.DemoPermissionsTableModel;
 import io.github.astrapi69.swing.table.model.GenericTableModel;
-import io.github.astrapi69.test.instances.TestPermissionFactory;
-import io.github.astrapi69.test.objects.Permission;
+import io.github.astrapi69.test.instance.TestPermissionFactory;
+import io.github.astrapi69.test.object.Permission;
 import io.github.astrapi69.window.adapter.CloseWindow;
 
 /**