+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGED:
 
 - update of gradle version to 7.3.3
 - update of gradle-plugin dependency com.github.ben-manes.versions.gradle.plugin to new version 0.41.0
+- update of dependency model-data version to 1.12
 - update of dependency swing-base-components to new version 1.1
 - update of dependency jobj-core to new version 5.2
 - moved all suffle panels to new package 'io.github.astrapi69.swing.table.panel.shuffle'