+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 ## Change log
 ----------------------
 
-Version 1.2-SNAPSHOT
+Version 2-SNAPSHOT
 -------------
 
+CHANGED:
 
+- update of JDK to newer version 11
+- update of gradle version to 7.5.1
+- update of gradle plugin dependency io.freefair.gradle:lombok-plugin to new version 6.5.1
+- update of gradle plugin dependency com.diffplug.spotless:spotless-plugin-gradle to new version 6.11.0
+- update of gradle-plugin dependency com.github.ben-manes.versions.gradle.plugin to new version 0.43.0
+- update of dependency silly-collection to new minor version 20.1
+- update of dependency jobj-core version to 7
+- update of dependency model-data to new minor module version 2.1
+- update of test dependency file-worker module version to 11.4
+- removed unused test dependency commons-text
+- update of test dependency junit-jupiter-api to new patch version 5.9.1
 
 Version 1.1
 -------------